- Add gif in splash screen - implemented ask permission for users

+++ app/src/main/java/net/grandpepper/caiena/grandpepper/util/AndroidSystemUtil.java
@@ -155,7 +155,7 @@ public class AndroidSystemUtil {
 
     public static void deleteDir() {
         File dir = getFolder();
-        if (dir.isDirectory()) {
+        if (dir.isDirectory() && dir.listFiles() != null) {
             for (File child : dir.listFiles())
                 child.delete();
         }
